+++ Network/API/Telegram/Bot/Object/Update/Message/Content/File.hs
@@ -7,6 +7,10 @@ import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Video as Expo
 import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Voice as Exports
 import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Caption as Exports
 import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Duration as Exports
+import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Filesize as Exports
+import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.MIME as Exports
+import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Height as Exports
+import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Width as Exports
 import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.URI as Exports
 
 import "aeson" Data.Aeson (FromJSON (parseJSON), withObject, (.:))