+++ Network/API/Telegram/Bot/Object/Update/Message/Content/File/Video.hs
@@ -2,7 +2,6 @@ module Network.API.Telegram.Bot.Object.Update.Message.Content.File.Video (Video 
 
 import "aeson" Data.Aeson (FromJSON (parseJSON), withObject, (.:), (.:?))
 import "base" Control.Applicative ((<*>))
-import "base" Data.Int (Int)
 import "base" Data.Maybe (Maybe)
 import "base" Data.Function (($))
 import "base" Data.Functor ((<$>))
@@ -11,10 +10,12 @@ import "base" Text.Show (Show)
 import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Duration (Duration)
 import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Filesize (Filesize)
 import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.MIME (MIME)
+import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Height (Height)
+import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Width (Width)
 import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.URI (URI)
 import Network.API.Telegram.Bot.Property (Identifiable (Identificator, ident))
 
-data Video = Video URI Duration Int Int (Maybe MIME) (Maybe Filesize) deriving Show
+data Video = Video URI Duration Height Width (Maybe MIME) (Maybe Filesize) deriving Show
 
 instance Identifiable Video where
 	type Identificator Video = URI