+++ CHANGELOG.md
@@ -147,4 +147,4 @@
 * Rename `Size` to `Photosize` and move its content to `Photo` module
 
 # 0.2.8
-* Define `Filesize` newtype in `Special` submodule to replace vague `Int`
+* Define `Filesize`, `Height` and `Width` newtypes in `Special` submodule to replace vague `Int`